requirements.txt
pickle==4.0.0
requests==2.28.1
pandas==1.5.2
google-api-python-client==2.124.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-storage==2.16.0